Refactored the multipart support to use the built-in encoders.
