+++ http-builder-ng-core/src/main/java/groovyx/net/http/MultipartContent.java
@@ -17,20 +17,10 @@ package groovyx.net.http;
 
 import groovy.lang.Closure;
 import groovy.lang.DelegatesTo;
-import groovyx.net.http.util.IoUtils;
 import org.apache.commons.lang3.RandomStringUtils;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Path;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.StringJoiner;
-import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 
 import static java.util.Collections.unmodifiableList;
@@ -49,10 +39,8 @@ import static java.util.Collections.unmodifiableList;
  *
  * which would define a `multipart/form-data` request with a field part and a file part with the specified properties.
  */
-public class MultipartContent implements ToServer {
+public class MultipartContent {
 
-    private static final String BOUNDARY_MARK = "--";
-    private static final String CRLF = "\r\n";
     private final List<MultipartPart> entries = new LinkedList<>();
     private final String boundary = RandomStringUtils.randomAlphanumeric(16);
 
@@ -82,89 +70,84 @@ public class MultipartContent implements ToServer {
         return content;
     }
 
+    /**
+     * Configures a field part with the given field name and value.
+     *
+     * @param fieldName the field name
+     * @param value the value
+     * @return a reference to this {@link MultipartContent} instance
+     */
     public MultipartContent field(String fieldName, String value) {
         return part(fieldName, value);
     }
 
+    /**
+     * Configures a field part with the given field name and value (of the specified content type).
+     *
+     * @param fieldName the field name
+     * @param contentType the value content type
+     * @param value the value
+     * @return a reference to this {@link MultipartContent} instance
+     */
     public MultipartContent field(String fieldName, String contentType, String value) {
         return part(fieldName, contentType, value);
     }
 
+    /**
+     * Configures a field part with the given field name and value.
+     *
+     * @param fieldName the field name
+     * @param value the value
+     * @return a reference to this {@link MultipartContent} instance
+     */
     public MultipartContent part(String fieldName, String value) {
         return part(fieldName, null, ContentTypes.TEXT.getAt(0), value);
     }
 
+    /**
+     * Configures a field part with the given field name and value (of the specified content type).
+     *
+     * @param fieldName the field name
+     * @param contentType the value content type
+     * @param value the value
+     * @return a reference to this {@link MultipartContent} instance
+     */
     public MultipartContent part(String fieldName, String contentType, String value) {
         return part(fieldName, null, contentType, value);
     }
 
-    // FIXME: should this be renamed MultipartRequestContent or is this the same for response ?
-    // FIXME: document
-    // FIXME: update guide documentation (and javadocs)
-    // FIXME: add interface over multipart to keep the DSL clean
-
+    /**
+     * Configures a file part with the specified properties. Encoders must be configured on the {@link HttpBuilder} to handle the content type
+     * of each configured part.
+     *
+     * @param fieldName the field name
+     * @param fileName the file name
+     * @param contentType the part content type
+     * @param content the part content (encoders must be configured)
+     * @return a reference to this {@link MultipartContent} instance
+     */
     public MultipartContent part(String fieldName, String fileName, String contentType, Object content) {
         entries.add(new MultipartPart(fieldName, fileName, contentType, content));
         return this;
     }
 
+    /**
+     * Iterates over the configured parts.
+     *
+     * @return an {@link Iterable} view of the configured parts.
+     */
     Iterable<MultipartPart> parts() {
         return unmodifiableList(entries);
     }
 
-    public String boundary() {
+    /**
+     * Used to retrieve the multipart content boundary used by this content. Encoders may provide their own boundary implementation or use
+     * the one provided by this method.
+     *
+     * @return a multipart boundary string for this content.
+     */
+    String boundary() {
         return boundary;
-    }
-
-    public InputStream toInputStream(final ChainedHttpConfig config) {
-        final ByteArrayOutputStream buffer = new ByteArrayOutputStream(1024);
-
-        try {
-            for (final MultipartPart part : entries) {
-                buffer.write(string(BOUNDARY_MARK, boundary(), CRLF));
-                buffer.write(string("Content-Type: ", part.getContentType(), CRLF));
-
-                if (part.getFileName() != null) {
-                    buffer.write(string("Content-Disposition: form-data; name=\"", part.getFieldName(), "\"; filename=\"", part.getFileName(), "\"", CRLF));
-
-                } else {
-                    buffer.write(string("Content-Disposition: form-data; name=\"", part.getFieldName(), "\"", CRLF));
-                }
-
-                //    BASE64('base64'), QUOTED_PRINTABLE('quoted-printable')
-//                buffer.write(string("Content-Transfer-Encoding: ", , CRLF));
-
-                buffer.write(string(CRLF));
-
-                final NestedToServer toServer = new NestedToServer();
-                BiConsumer<ChainedHttpConfig, ToServer> encoder = config.findEncoder(part.getContentType());
-                encoder.accept(new PartConfig(config, part), toServer);
-                buffer.write(toServer.getBytes());
-
-                buffer.write(string(CRLF));
-            }
-
-            buffer.write(string(BOUNDARY_MARK, boundary(), BOUNDARY_MARK));
-
-        } catch (IOException e) {
-            // FIXME: do better?
-            e.printStackTrace();
-        }
-
-        return new ByteArrayInputStream(buffer.toByteArray());
-    }
-
-    private static byte[] string(final String... str) {
-        final StringJoiner joiner = new StringJoiner("");
-        for (final String s : str) {
-            joiner.add(s);
-        }
-        return joiner.toString().getBytes(StandardCharsets.UTF_8);
-    }
-
-    @Override
-    public void toServer(final InputStream inputStream) {
-
     }
 
     /**
@@ -199,74 +182,5 @@ public class MultipartContent implements ToServer {
         public Object getContent() {
             return content;
         }
-    }
-}
-
-class NestedToServer implements ToServer {
-
-    private byte[] bytes;
-
-    @Override
-    public void toServer(final InputStream inputStream) {
-        try {
-            bytes = IoUtils.streamToBytes(inputStream);
-        } catch (IOException e) {
-            // FIXME: something better?
-            e.printStackTrace();
-        }
-    }
-
-    public byte[] getBytes() {
-        return bytes;
-    }
-}
-
-class PartConfig implements ChainedHttpConfig {
-
-    private final ChainedHttpConfig config;
-    private final MultipartContent.MultipartPart part;
-
-    PartConfig(final ChainedHttpConfig config, final MultipartContent.MultipartPart part) {
-        this.config = config;
-        this.part = part;
-    }
-
-    @Override
-    public Map<Map.Entry<String, Object>, Object> getContextMap() {
-        return config.getContextMap();
-    }
-
-    @Override
-    public ChainedResponse getChainedResponse() {
-        return config.getChainedResponse();
-    }
-
-    @Override
-    public ChainedRequest getChainedRequest() {
-        return (ChainedRequest) getRequest();
-    }
-
-    @Override
-    public ChainedHttpConfig getParent() {
-        return null;
-    }
-
-    @Override
-    public void context(String contentType, Object id, Object obj) {
-        config.context(contentType, id, obj);
-    }
-
-    @Override
-    public Request getRequest() {
-        final HttpConfigs.BasicRequest request = new HttpConfigs.BasicRequest(null);
-        request.setContentType(part.getContentType());
-        request.setCharset(StandardCharsets.UTF_8); // TODO: ok?
-        request.setBody(part.getContent());
-        return request;
-    }
-
-    @Override
-    public Response getResponse() {
-        return config.getResponse();
     }
 }
